Implement issue #540: rename Preferences to Settings, show Dock icon while open
Rename:
- PreferencesView -> SettingsView (file + type)
- PreferencesWindowController -> SettingsWindowController (file + type)
- Window title: 'ControlPlane' -> 'ControlPlane Settings'
- Menu item: 'Preferences...' -> 'Settings...' (Cmd+, shortcut unchanged)
- Selector: openPreferences -> openSettings

Dock icon / Cmd+Tab:
- SettingsWindowController.show() calls NSApp.setActivationPolicy(.regular)
  before makeKeyAndOrderFront so the Dock icon and Cmd+Tab entry are
  present from the very first visible frame
- windowWillClose() reverts to NSApp.setActivationPolicy(.accessory) so
  the app disappears from the Dock and switcher once Settings is closed
- LSUIElement in Info.plist is unchanged; the policy is toggled at runtime
